Add files via upload
My projects and reflection word document.
Briefly summarizing The Gaming Room client and their software requirements, the client, in this case, was The Gaming Room, and they sought the enhancement and expansion of their existing game, "Draw It or Lost It," originally designed for Android. The objective was to improve software functionality and extend its availability to other platforms. The documentation was crafted to be clear, concise, and focused on key points.

Regarding the process of working through a design document, it proved beneficial for maintaining focus on the task at hand, prioritizing important tasks, and discerning those that required less time and attention.

If there was one part of the work on these documents to revise, it would be the "Recommendations" section, acknowledging the challenge in articulating the necessary details and justifications.

Interpreting the user's needs involved breaking down the game's functionality to streamline the process of expanding to different operating platforms while maintaining a consistent user experience. Emphasizing the importance of user needs in the design process is crucial since users ultimately control the product's success; their trust and satisfaction are paramount.

The approach to designing software centered on clear documentation and the strategic separation of game functionality, facilitating easier updates and maintenance. For future software applications, the focus would remain on these principles to ensure efficient analysis, design, and implementation.